Move GlimmerX `template()` to root of DSL
This is pretty non-obvious and we should make sure to capture it in the
forthcoming `ARCHITECTURE.md`!
