+++ packages/environment-glimmerx/-private/dsl/index.d.ts
@@ -23,12 +23,17 @@ import './integration-declarations';
 
 import { ResolveOrReturn } from '@glint/template/-private/dsl';
 import {
+  AcceptsBlocks,
+  AnyContext,
+  AnyFunction,
   BoundModifier,
   DirectInvokable,
   EmptyObject,
+  HasContext,
   Invokable,
   Invoke,
   InvokeDirect,
+  TemplateContext,
 } from '@glint/template/-private/integration';
 
 export declare function resolve<T extends DirectInvokable>(item: T): T[typeof InvokeDirect];
@@ -49,3 +54,17 @@ export declare function resolve<Args extends unknown[], T>(
 ): (named: EmptyObject, ...args: Args) => T;
 
 export declare const resolveOrReturn: ResolveOrReturn<typeof resolve>;
+
+// We customize the top-level `template` wrapper function for this environment to
+// return a type that's assignable to `TemplateComponent` from '@glimmerx/component'.
+// Longer term we should rationalize this to a type that doesn't carry extra baggage
+// and likely comes from a more sensible path.
+
+import { TemplateComponent } from '@glimmerx/component';
+
+export declare function template<
+  Signature extends AnyFunction = (args: EmptyObject) => AcceptsBlocks<EmptyObject>,
+  Context extends AnyContext = TemplateContext<void, EmptyObject, EmptyObject, void>
+>(
+  f: (𝚪: Context, χ: never) => void
+): TemplateComponent<never> & (new () => Invokable<Signature> & HasContext<Context>);
